Query device only once
Currently every invocation of `solve_cuda_batch` function checks
the kind of GPU to determine the block size.

This is wasteful, and we observe the case where this takes a lot
more than the computation itself.

We can workaround this by caching the result.
The proposed change take advantage of static storage, so that it
is initialized only once.

This works the same way as the current behavior, for cases where
one process is assigned for each GPU.
(which I think is almost always the case, when using PyTorch.)

If there should be a case where one process is assigned to use
multiple different kind of GPUs, then this might not give the
indended value.
But such usecase would require fully customized PyTorch distributed
trainign framework, so I assume it's highly unlikely.